+++ plugins/ruby-grape-rails/hooks/scripts/inject-iron-laws.sh
@@ -3,7 +3,7 @@ set -o nounset
 set -o pipefail
 
 # GENERATED FROM iron-laws.yml — DO NOT EDIT
-# Last generated: 2026-03-22T22:17:20Z
+# Last generated: 2026-03-23T07:12:47Z
 
 command -v jq >/dev/null 2>&1 || exit 0